# openapi.yaml
openapi: 3.0.3
info:
  title: Esketit Music Server API
  version: 1.3.0
  description: |
    REST API for listing and uploading media files, managing albums, tracks, authors,
    user playlists, and authenticating users with access and refresh tokens.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/tracks:
    get:
      tags: [Tracks]
      summary: List tracks
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
          description: Page number, defaults to 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
          description: Page size, defaults to 20
        - in: query
          name: query
          schema:
            type: string
          description: Case-insensitive substring match against track name
        - in: query
          name: authorId
          schema:
            type: integer
            minimum: 1
          description: Filter tracks by author membership
        - in: query
          name: albumId
          schema:
            type: integer
            minimum: 1
          description: Filter tracks by album ID
      responses:
        '200':
          description: Paginated track list sorted by ID asc
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTracks'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [Tracks]
      summary: Create track
      description: |
        Create a track record that points to an already uploaded song file via
        `audioFilePath`, typically `/api/songs/{name}`.

        This endpoint only creates track metadata. It does not upload an MP3, move
        an MP3, or delete any MP3 file. The client must upload the song separately
        before creating the track.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertTrackRequest'
      responses:
        '201':
          description: Created track
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/PlainError'
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/PlainError'
    Forbidden:
      description: Insufficient permissions
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/PlainError'
...
    InternalServerError:
      description: Internal server error
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/PlainError'
  schemas:
    PlainError:
      type: string
...
    AdditionalInfo:
      type: object
      required: [type]
      properties:
        type:
          type: string
      additionalProperties: true
...
    SourceMetadata:
      type: object
      properties:
        provider:
          type: string
        kind:
          type: string
        identity:
          type: object
          additionalProperties: true
        url:
          type: string
          format: uri
      required: [provider, identity]
      additionalProperties: true
...
    Track:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        authorIds:
          type: array
          items:
            type: integer
            format: int64
            minimum: 1
        albumId:
          type: integer
          format: int64
          minimum: 1
        audioFilePath:
          type: string
        additionalInfo:
          type: array
          items:
            $ref: '#/components/schemas/AdditionalInfo'
        sourceMetadata:
          type: array
          items:
            $ref: '#/components/schemas/SourceMetadata'
        isFavorite:
          type: boolean
        isAvailable:
          type: boolean
      required: [id, name, authorIds, albumId, audioFilePath, additionalInfo, sourceMetadata, isFavorite, isAvailable]
    UpsertTrackRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        authorIds:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
            minimum: 1
        albumId:
          type: integer
          format: int64
          minimum: 1
        albumOrder:
          type: integer
          minimum: 0
        audioFilePath:
          type: string
          description: |
            Path or URL of an existing song file, typically `/api/songs/{name}`.

            This is a reference to an existing uploaded MP3, not an owned file
            object. Changing this field only changes the stored reference.
            Deleting the track does not delete the referenced MP3 from storage.
          minLength: 1
        additionalInfo:
          type: array
          items:
            $ref: '#/components/schemas/AdditionalInfo'
        sourceMetadata:
          type: array
          items:
            $ref: '#/components/schemas/SourceMetadata'
      required: [name, authorIds, albumId, albumOrder, audioFilePath, additionalInfo, sourceMetadata]
...
    PaginatedTracks:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Track'
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
      required: [items, page, pageSize, totalItems, totalPages]
...